Rename fileserver to more generic cook.sidecar
