+++ sidecar/.dockerignore
@@ -1,0 +1,7 @@
+build
+dist
+*.egg-info
+*.egg/
+*.pyc
+*.swp
+__pycache__
