+++ sidecar/README.md
@@ -18,13 +18,13 @@ Usage:
 
 The `COOK_WORKDIR` environment variable must be set. Only files with `COOK_WORKDIR` as the root will be served.
 
-```fileserver PORT [NUM_WORKERS]```
+```cook-sidecar PORT [NUM_WORKERS]```
 
 `NUM_WORKERS` defaults to `4`.
 
 Examples:
 
 ```bash
-$ fileserver 8000
-$ fileserver 8000 4
+$ cook-sidecar 8000
+$ cook-sidecar 8000 4
 ```
