+++ sidecar/cook/sidecar/version.py
